+++ src/components/extension/index.ts
@@ -1,0 +1,2 @@
+export { ShareIntentRoot } from './ShareIntentRoot';
+export { IOSShareIntentRoot } from './IOSShareIntentRoot';
